[Pytorch Mobile] Add flag to module in save_for_mobile indicating its a mobile model
Add a flag so when .ptls are loaded as full jit we can know they are .ptls.

This can help warn users of potential logic errors like in export_opnames. There is also already precedence for silently adding a flag like this to the original jit in optimize_for_mobile https://fburl.com/diffusion/cdfv2100

Differential Revision: [D26617770](https://our.internmc.facebook.com/intern/diff/D26617770/)

ghstack-source-id: 122345508
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53034