[package] Add `deny` method to `PackageExporter`
**Summary**
This commit adds a `deny` method to `PackageExporter` that allows
modules to be prohibited during the packaging process. A dependency on a
module matching the names or globs that `deny` was called with will
cause an exception to be raised.

**Test Plan**
This commit adds unit tests to `PackagingTest` for this new method:
`test_deny` and `test_deny_glob`.

**Fixes**
This commit fixes #53217.

ghstack-source-id: 633fa5407670acb167d809a1ad05a47d9bc1f959
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53233